Provide boilerplate solution
