+++ exercises/concept/dancing-dots/lib/dancing_dots/animation.ex
@@ -1,47 +1,16 @@
-defmodule DancingDots.Animation do
+ defmodule DancingDots.Animation do
   @type dot :: DancingDots.Dot.t()
-  @type opts :: any
+  @type opts :: keyword
   @type error :: any
   @type frame_number :: pos_integer
 
-  @callback init(opts :: opts) :: {:ok, opts} | {:error, error}
-  @callback handle_frame(dot :: dot, n :: frame_number, opts :: opts) :: dot
+  # Please implement the module
+ end
 
-  defmacro __using__(_) do
-    quote do
-      @behaviour DancingDots.Animation
-      def init(opts), do: {:ok, opts}
-      defoverridable init: 1
-    end
-  end
-end
+ defmodule DancingDots.Flicker do
+  # Please implement the module
+ end
 
-defmodule DancingDots.Flicker do
-  use DancingDots.Animation
-
-  @impl DancingDots.Animation
-  def handle_frame(dot, frame_number, _opts) do
-    opacity = if rem(frame_number, 4) == 0, do: dot.opacity / 2, else: dot.opacity
-    %{dot | opacity: opacity}
-  end
-end
-
-defmodule DancingDots.Zoom do
-  use DancingDots.Animation
-
-  @impl DancingDots.Animation
-  def init(opts) do
-    velocity = Keyword.get(opts, :velocity)
-
-    if is_number(velocity) do
-      {:ok, [velocity: velocity]}
-    else
-      {:error, "Expected required option :velocity to be a number, got: #{inspect(velocity)}"}
-    end
-  end
-
-  @impl DancingDots.Animation
-  def handle_frame(dot, frame_number, opts) do
-    %{dot | radius: dot.radius + opts[:velocity] * (frame_number - 1)}
-  end
-end
+ defmodule DancingDots.Zoom do
+  # Please implement the module
+ end
